Add control parameter framework.
All EMOA implementations should pass their control list() through the
appropriate *_control() functions to validate the parameters or fill in
missing values with default values. Because all parameters are now
stored in the control list(), the default logger hat to be adapted to
look in the control list for a reference point instead of in the
environment of the EMOA function.