Implement automatic versioning.
This commit adds an automatic versioning scheme to the Makefile. If this
is a clean working copy, we count the number of commits since the last
tagged minor release (tag name 'v<major>.<minor>') and then use
'<major>.<minor>-<#commits>' as the version number. If the working copy
is not clean, we again determine the major and minor version from the
last tagged commit and then use the number of seconds since the
Beginning of Time(tm) as the revision. This ensures, that
dirty/development packages always have a higher revision number than
checked in releases.