Add a completely unrealistic benchmark example.
This example compares the nsga2 implementation in mco with the sms_emoa
implementation from the examples directory of the emoa package based on
50 runs on the ZDT3 test function. The observed performance metrics
should not be used to draw any conclusions about the relativ merit of
either algorithm.

The main aim of this example is to provide a simple benchmark to quickly
check if the performance of the sms_emoa implementation is degrading
over time and if any changes in the emoa package speed up the
performance (in cpu time) of the algorithm.