Fix another bug in hypervolume_contribution().
For more than two dimensions, hypervolume_contribution() permuted the
columns of the matrix passed in. This resulted from a missing
duplicated() in do_hv_contrib in hypervolume.c.

USER VISIBLE CHANGE: sms_emoa() should finally work on three or more
  dimensional problems.

RATIONAL: Bugfix.

WHAT TO CHANGE: Nothing. But if you've used a previous version of the
  package for many objective optimization, you should check your results
  and rerun the optimization.